requirements.txt
Django>=4.2.0
psycopg2-binary>=2.9.0
django-redis>=5.2.0
celery>=5.3.0
redis>=4.5.0
python-dotenv>=1.0.0
djangorestframework>=3.14.0
opensearch-py>=2.4.0
kafka-python>=2.0.2
elasticsearch>=8.0.0
requests>=2.31.0
drf-spectacular>=0.26.0
pytest>=7.4.0
pytest-django>=4.5.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
factory-boy>=3.3.0
faker>=19.0.0
responses>=0.23.0
freezegun>=1.2.0
